+++ brownie/project/compiler/solidity.py
@@ -1,7 +1,6 @@
 #!/usr/bin/python3
 
 import logging
-import re
 from collections import deque
 from hashlib import sha1
 from typing import Any, Dict, List, Optional, Set, Tuple
@@ -9,11 +8,11 @@ from typing import Any, Dict, List, Optional, Set, Tuple
 import solcast
 import solcx
 from requests.exceptions import ConnectionError
-from semantic_version import NpmSpec, Version
+from semantic_version import Version
 from solcast.nodes import NodeBase
 
 from brownie._config import EVM_EQUIVALENTS
-from brownie.exceptions import CompilerError, IncompatibleSolcVersion, PragmaError
+from brownie.exceptions import CompilerError, IncompatibleSolcVersion
 from brownie.project.compiler.utils import expand_source_map
 
 from . import sources
@@ -25,7 +24,6 @@ sh.setLevel(10)
 sh.setFormatter(logging.Formatter("%(message)s"))
 solcx_logger.addHandler(sh)
 
-PRAGMA_REGEX = re.compile(r"pragma +solidity([^;]*);")
 AVAILABLE_SOLC_VERSIONS = None
 
 
@@ -94,6 +92,14 @@ def install_solc(*versions: str) -> None:
         solcx.install_solc(str(version), show_progress=True)
 
 
+def get_abi(contract_source: str) -> Dict:
+    """Given a contract source, returns a dict of {name: abi}"""
+    version = find_best_solc_version({"<stdin>": contract_source})
+    set_solc_version(version)
+    compiled = solcx.compile_source(contract_source, allow_empty=True, output_values=["abi"])
+    return {k.rsplit(":")[-1]: v["abi"] for k, v in compiled.items()}
+
+
 def find_solc_versions(
     contract_sources: Dict[str, str],
     install_needed: bool = False,
@@ -122,16 +128,12 @@ def find_solc_versions(
     new_versions = set()
 
     for path, source in contract_sources.items():
-
-        pragma_string = next(PRAGMA_REGEX.finditer(source), None)
-        if pragma_string is None:
-            raise PragmaError(f"No version pragma in '{path}'")
-        pragma_specs[path] = NpmSpec(pragma_string.groups()[0])
+        pragma_specs[path] = sources.get_pragma_spec(source, path)
         version = pragma_specs[path].select(installed_versions)
 
         if not version and not (install_needed or install_latest):
             raise IncompatibleSolcVersion(
-                f"No installed solc version matching '{pragma_string[0]}' in '{path}'"
+                f"No installed solc version matching '{pragma_specs[path]}' in '{path}'"
             )
 
         # if no installed version of solc matches the pragma, find the latest available version
@@ -139,7 +141,7 @@ def find_solc_versions(
 
         if not version and not latest:
             raise IncompatibleSolcVersion(
-                f"No installable solc version matching '{pragma_string[0]}' in '{path}'"
+                f"No installable solc version matching '{pragma_specs[path]}' in '{path}'"
             )
 
         if not version or (install_latest and latest > version):
@@ -191,10 +193,7 @@ def find_best_solc_version(
 
     for path, source in contract_sources.items():
 
-        pragma_string = next(PRAGMA_REGEX.finditer(source), None)
-        if pragma_string is None:
-            raise PragmaError(f"No version pragma in '{path}'")
-        pragma_spec = NpmSpec(pragma_string.groups()[0])
+        pragma_spec = sources.get_pragma_spec(source, path)
         installed_versions = [i for i in installed_versions if i in pragma_spec]
         available_versions = [i for i in available_versions if i in pragma_spec]
 
@@ -249,7 +248,7 @@ def _get_unique_build_json(
     return {
         "allSourcePaths": paths,
         "bytecode": bytecode,
-        "bytecodeSha1": sha1(bytecode[:-68].encode()).hexdigest(),
+        "bytecodeSha1": sha1(_remove_metadata(bytecode).encode()).hexdigest(),
         "coverageMap": {"statements": statement_map, "branches": branch_map},
         "dependencies": [i.name for i in contract_node.dependencies],
         "offset": contract_node.offset,
@@ -258,7 +257,7 @@ def _get_unique_build_json(
     }
 
 
-def _format_link_references(evm: Dict) -> Dict:
+def _format_link_references(evm: Dict) -> str:
     # Standardizes formatting for unlinked libraries within bytecode
     bytecode = evm["bytecode"]["object"]
     references = [
@@ -267,6 +266,13 @@ def _format_link_references(evm: Dict) -> Dict:
     for n, loc in [(i[0], x["start"] * 2) for i in references for x in i[1]]:
         bytecode = f"{bytecode[:loc]}__{n[:36]:_<36}__{bytecode[loc+40:]}"
     return bytecode
+
+
+def _remove_metadata(bytecode: str) -> str:
+    if not bytecode:
+        return ""
+    idx = -(int(bytecode[-4:], 16) + 2) * 2
+    return bytecode[:idx]
 
 
 def _generate_coverage_data(