+++ brownie/project/main.py
@@ -9,6 +9,7 @@ from pathlib import Path
 from typing import Any, Dict, Iterator, KeysView, List, Optional, Set, Union
 
 import requests
+from semantic_version import Version
 from tqdm import tqdm
 
 from brownie._config import (
@@ -26,14 +27,15 @@ from brownie.network.state import _remove_contract
 from brownie.project import compiler
 from brownie.project.build import BUILD_KEYS, Build
 from brownie.project.ethpm import get_deployment_addresses, get_manifest
-from brownie.project.sources import Sources, get_hash
+from brownie.project.sources import Sources, get_hash, get_pragma_spec
 from brownie.utils import color
 
 FOLDERS = [
     "contracts",
+    "interfaces",
     "scripts",
-    "reports",
     "tests",
+    "reports",
     "build",
     "build/contracts",
     "build/deployments",
@@ -44,19 +46,18 @@ _loaded_projects = []
 
 
 class _ProjectBase:
-    def __init__(self, name: str, contract_sources: Dict, project_path: Optional[Path]) -> None:
-        self._path = project_path
-        self._name = name
-        self._sources = Sources(contract_sources)
-        self._build = Build(self._sources)
 
-    def _compile(self, sources: Dict, compiler_config: Dict, silent: bool) -> None:
+    _path: Optional[Path]
+    _sources: Sources
+    _build: Build
+
+    def _compile(self, contract_sources: Dict, compiler_config: Dict, silent: bool) -> None:
         allow_paths = None
         if self._path is not None:
             allow_paths = self._path.joinpath("contracts").as_posix()
         compiler_config.setdefault("solc", {})
         build_json = compiler.compile_and_format(
-            sources,
+            contract_sources,
             solc_version=compiler_config["solc"].get("version", None),
             optimize=compiler_config["solc"].get("optimize", None),
             runs=compiler_config["solc"].get("runs", None),
@@ -64,6 +65,7 @@ class _ProjectBase:
             minify=compiler_config["minify_source"],
             silent=silent,
             allow_paths=allow_paths,
+            interface_sources=self._sources.get_interface_sources(),
         )
         for data in build_json.values():
             if self._path is not None:
@@ -114,19 +116,24 @@ class Project(_ProjectBase):
     """
 
     def __init__(self, name: str, project_path: Path) -> None:
-        contract_sources: Dict = {}
-        for path in project_path.glob("contracts/**/*"):
-            if "/_" in path.as_posix() or path.suffix not in (".sol", ".vy"):
-                continue
-            with path.open() as fp:
-                source = fp.read()
-            path_str: str = path.relative_to(project_path).as_posix()
-            contract_sources[path_str] = source
-        super().__init__(name, contract_sources, project_path)
         self._path: Path = project_path
+        self._name = name
+        self._active = False
+        self.load()
+
+    def load(self) -> None:
+        """Compiles the project contracts, creates ContractContainer objects and
+        populates the namespace."""
+        if self._active:
+            raise ProjectAlreadyLoaded("Project is already active")
+
+        contract_sources = _load_sources(self._path, "contracts", False)
+        interface_sources = _load_sources(self._path, "interfaces", True)
+        self._sources = Sources(contract_sources, interface_sources)
+        self._build = Build(self._sources)
 
         contract_list = self._sources.get_contract_list()
-        for path in list(project_path.glob("build/contracts/*.json")):
+        for path in list(self._path.glob("build/contracts/*.json")):
             try:
                 with path.open() as fp:
                     build_json = json.load(fp)
@@ -137,20 +144,12 @@ class Project(_ProjectBase):
                 continue
             self._build._add(build_json)
 
-        self._active = False
-        self.load()
-
-    def load(self) -> None:
-        """Compiles the project contracts, creates ContractContainer objects and
-        populates the namespace."""
-        if self._active:
-            raise ProjectAlreadyLoaded("Project is already active")
-
         self._compiler_config = _load_project_compiler_config(self._path)
 
         # compile updated sources, update build
         changed = self._get_changed_contracts()
         self._compile(changed, self._compiler_config, False)
+        self._save_interface_hashes()
         self._create_containers()
         self._load_deployments()
 
@@ -169,32 +168,60 @@ class Project(_ProjectBase):
         _loaded_projects.append(self)
 
     def _get_changed_contracts(self) -> Dict:
-        changed = [i for i in self._sources.get_contract_list() if self._compare_build_json(i)]
-        final = set(changed)
-        for contract_name in changed:
+        # get list of changed interfaces and contracts
+        old_hashes = self._load_interface_hashes()
+        new_hashes = self._sources.get_interface_hashes()
+        interfaces = [k for k, v in new_hashes.items() if old_hashes.get(k, None) != v]
+        contracts = [i for i in self._sources.get_contract_list() if self._compare_build_json(i)]
+
+        # get dependents of changed sources
+        final = set(contracts + interfaces)
+        for contract_name in list(final):
             final.update(self._build.get_dependents(contract_name))
+
+        # remove outdated build artifacts
         for name in [i for i in final if self._build.contains(i)]:
             self._build._remove(name)
+
+        # get final list of changed source paths
+        final.difference_update(interfaces)
         changed_set: Set = set(self._sources.get_source_path(i) for i in final)
-        return dict((i, self._sources.get(i)) for i in changed_set)
+        return {i: self._sources.get(i) for i in changed_set}
+
+    def _load_interface_hashes(self) -> Dict:
+        try:
+            with self._path.joinpath("build/interfaces.json").open() as fp:
+                return json.load(fp)
+        except (FileNotFoundError, json.JSONDecodeError):
+            return {}
+
+    def _save_interface_hashes(self) -> None:
+        interface_hashes = self._sources.get_interface_hashes()
+        with self._path.joinpath("build/interfaces.json").open("w") as fp:
+            json.dump(interface_hashes, fp, sort_keys=True, indent=2)
 
     def _compare_build_json(self, contract_name: str) -> bool:
         config = self._compiler_config
+        # confirm that this contract was previously compiled
         try:
             source = self._sources.get(contract_name)
             build_json = self._build.get(contract_name)
         except KeyError:
             return True
-        if build_json["sha1"] != get_hash(
-            source, contract_name, config["minify_source"], build_json["language"]
-        ):
+        # compare source hashes
+        hash_ = get_hash(source, contract_name, config["minify_source"], build_json["language"])
+        if build_json["sha1"] != hash_:
             return True
+        # compare compiler settings
         if _compare_settings(config, build_json["compiler"]):
             return True
-        if build_json["language"] == "Solidity" and _compare_settings(
-            config["solc"], build_json["compiler"]
-        ):
-            return True
+        if build_json["language"] == "Solidity":
+            # compare solc-specific compiler settings
+            if _compare_settings(config["solc"], build_json["compiler"]):
+                return True
+            # compare solc pragma against compiled version
+            if Version(build_json["compiler"]["version"]) not in get_pragma_spec(source):
+                return True
         return False
 
     def _load_deployments(self) -> None:
@@ -273,7 +300,10 @@ class TempProject(_ProjectBase):
     compiled via project.compile_source"""
 
     def __init__(self, name: str, contract_sources: Dict, compiler_config: Dict) -> None:
-        super().__init__(name, contract_sources, None)
+        self._path = None
+        self._name = name
+        self._sources = Sources(contract_sources, {})
+        self._build = Build(self._sources)
         self._compile(contract_sources, compiler_config, True)
         self._create_containers()
 
@@ -463,3 +493,15 @@ def _compare_settings(left: Dict, right: Dict) -> bool:
     return next(
         (True for k, v in left.items() if v and not isinstance(v, dict) and v != right[k]), False
     )
+
+
+def _load_sources(project_path: Path, subfolder: str, allow_json: bool) -> Dict:
+    contract_sources: Dict = {}
+    for path in project_path.glob(f"{subfolder}/**/*"):
+        if "/_" in path.as_posix() or path.suffix not in (".sol", ".vy", ".json"):
+            continue
+        with path.open() as fp:
+            source = fp.read()
+        path_str: str = path.relative_to(project_path).as_posix()
+        contract_sources[path_str] = source
+    return contract_sources