+++ CHANGELOG.md
@@ -7,7 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/iamdefinitelyahuman/brownie)
 ### Added
+- `interfaces/` folder for interface sources
+- tested support for Vyper with ethPM
 - Progress bar when downloading a Brownie mix
+- `get_abi` method for Solidity and Vyper compilers
+
+### Changed
+- Move check for new contract sources from `Project.__init__` to `Project.load`
+
+### Fixed
+- Check pragma statements when determining if a contract should be recompiled
+- Understand `abstract contract` when regexing contract source
 
 ## [1.4.2](https://github.com/iamdefinitelyahuman/brownie/tree/v1.4.2) - 2020-01-10
 ### Added